chore: move SignatoryAndAddress struct to wire package
